+++ hutool-core/src/test/java/com/xiaoleilu/hutool/core/util/StrUtilTest.java
@@ -1,5 +1,7 @@
 package com.xiaoleilu.hutool.core.util;
 
+import java.util.List;
+
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -39,5 +41,17 @@ public class StrUtilTest {
 		String str = "aaabbbcccdddaadfdfsdfsdf0";
 		String[] cut = StrUtil.cut(str, 4);
 		Console.log(cut);
+	}
+	
+	@Test
+	public void splitTest(){
+		String str = "a,b ,c,d,,e";
+		List<String> split = StrUtil.split(str, ',', -1, true, true);
+		//测试空是否被去掉
+		Assert.assertEquals(5, split.size());
+		//测试去掉两边空白符是否生效
+		Assert.assertEquals("b", split.get(1));
+		
+		
 	}
 }